requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.5
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==4.0.2
django-debug-toolbar==4.3.0
django-webpush==0.3.4
future==0.18.2
google-api-core==1.31.2
google-auth==1.35.0
google-crc32c==1.1.2
googleapis-common-protos==1.53.0
grpcio==1.44.0
html5lib==1.1
http-ece==1.1.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
mysqlclient==2.1.0
packaging==21.0
Pillow==9.0.1
proto-plus==1.19.0
protobuf==3.18.0
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
pywebpush==1.9.4
qrcode==7.3.1
reportlab==3.6.6
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.6
webencodings==0.5.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
